requirements.txt
google.cloud
io
es
enum
PIL
opencv
numpy
six
win32gui
win32api
win32con
win32com.client
win32ui